replace solve-linear with solve-linear-v2
